+++ src/api/mock_gatewayer_test.go
@@ -93,6 +93,20 @@ func (_m *MockGatewayer) DecryptWallet(wltID string, password []byte) (*wallet.W
 	return r0, r1
 }
 
+// Disconnect provides a mock function with given fields: id
+func (_m *MockGatewayer) Disconnect(id uint64) error {
+	ret := _m.Called(id)
+
+	var r0 error
+	if rf, ok := ret.Get(0).(func(uint64) error); ok {
+		r0 = rf(id)
+	} else {
+		r0 = ret.Error(0)
+	}
+
+	return r0
+}
+
 // EncryptWallet provides a mock function with given fields: wltID, password
 func (_m *MockGatewayer) EncryptWallet(wltID string, password []byte) (*wallet.Wallet, error) {
 	ret := _m.Called(wltID, password)
@@ -459,29 +473,6 @@ func (_m *MockGatewayer) GetHealth() (*daemon.Health, error) {
 	} else {
 		if ret.Get(0) != nil {
 			r0 = ret.Get(0).(*daemon.Health)
-		}
-	}
-
-	var r1 error
-	if rf, ok := ret.Get(1).(func() error); ok {
-		r1 = rf()
-	} else {
-		r1 = ret.Error(1)
-	}
-
-	return r0, r1
-}
-
-// GetIncomingConnections provides a mock function with given fields:
-func (_m *MockGatewayer) GetIncomingConnections() ([]daemon.Connection, error) {
-	ret := _m.Called()
-
-	var r0 []daemon.Connection
-	if rf, ok := ret.Get(0).(func() []daemon.Connection); ok {
-		r0 = rf()
-	} else {
-		if ret.Get(0) != nil {
-			r0 = ret.Get(0).([]daemon.Connection)
 		}
 	}